delete user functionality, export popup design, language field required

+++ app/Http/Controllers/ProfileController.php
@@ -41,23 +41,34 @@ class ProfileController extends Controller
     /**
      * Delete the user's account.
      */
-    public function destroy(Request $request): RedirectResponse
+    public function destroy($id, Request $request): RedirectResponse
     {
-        $request->validateWithBag('userDeletion', [
-            'password' => ['required', 'current_password'],
-        ]);
+        // $request->validateWithBag('userDeletion', [
+        //     'password' => ['required', 'current_password'],
+        // ]);
+        // Log::info('Delete request received for user ID: ' . $id);
+        $user=User::findOrFail($id);
+        if(auth()->user()->id === $user->id){
+            return redirect()->route('users.index')->with('error','You cannot delete your own account');
+        }
+        // $user = $request->user();
 
-        $user = $request->user();
-
-        Auth::logout();
+        // Auth::logout();
 
         $user->delete();
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
+        // $request->session()->invalidate();
+        // $request->session()->regenerateToken();
 
-        return Redirect::to('/');
+        // return Redirect::to('/');
+        return redirect()->route('users.index')->with('success','User deleted successfully!');
     }
+    // public function destroyUser(Request $request):RedirectResponse
+    // {
+    //     $user->delete();
+    //     return redirect()->route('users.index')->with('success','User deleted successfully!');
+
+    // }
 
     public function users()
     {
